Update and rename Sorting techniques to Searching techniques
Demonstrate the time difference between Binary and Linear/Sequential Search with some sample data. The end result shows that a linear search is much more efficient if we are using a linked list. Although, the complexity of the linear list is O(N) which is more than that of Binary search; O(log(N)).